текущие значения завершены, нужны тесты под нагрузкой

+++ WTB_11/DataStructures/VPBDateTimeTemp.cs
@@ -10,12 +10,12 @@ namespace WPB_11.DataStructures
     {
         public struct VPBDateTimeTempStruct
         {
-            public byte Hour;
-            public byte Minute;
-            public byte Second;
-            public byte Date;
-            public byte Month;
-            public byte Year;
+            public int Hour;
+            public int Minute;
+            public int Second;
+            public int Date;
+            public int Month;
+            public int Year;
             public byte Temperature_H;
             public byte Temperature_L;
         }
